MixedModelsSim.jl

MixedModelsSim.cyclicshift

MixedModelsSim.factorproduct

MixedModelsSim.nlevels

MixedModelsSim.pooled!

MixedModelsSim.power_table

MixedModelsSim.replicate

MixedModelsSim.sim_to_df

MixedModelsSim.simdat_crossed

MixedModelsSim.simulate_waldtests

MixedModelsSim.withinitem

MixedModelsSim.cyclicshift — Method

cyclicshift(v::AbstractVector, nrow)

Return an eltype(v) matrix of size nrow by length(v) with each column consisting of v followed by a cyclic shift of v followed by ...

The return value is used to counterbalance levels of conditions in withinitem.

cyclicshift('a':'d', 8)

MixedModelsSim.factorproduct — Method

factorproduct(facs...)

Return a Vector{NamedTuple} obtained by crossing facs....

The arguments should be coerceable to a Tables.RowTable with rowtable.

The value is a Tables.RowTable and hence can be converted to a DataFrame.

Example

julia> DataFrame(factorproduct((item=nlevels(3,'I'),), (subj=nlevels(5), age=["Y","Y","Y","O","O"])))
15×3 DataFrame
│ Row │ item   │ subj   │ age    │
│     │ String │ String │ String │
├─────┼────────┼────────┼────────┤
│ 1   │ I1     │ S1     │ Y      │
│ 2   │ I2     │ S1     │ Y      │
│ 3   │ I3     │ S1     │ Y      │
│ 4   │ I1     │ S2     │ Y      │
│ 5   │ I2     │ S2     │ Y      │
│ 6   │ I3     │ S2     │ Y      │
│ 7   │ I1     │ S3     │ Y      │
│ 8   │ I2     │ S3     │ Y      │
│ 9   │ I3     │ S3     │ Y      │
│ 10  │ I1     │ S4     │ O      │
│ 11  │ I2     │ S4     │ O      │
│ 12  │ I3     │ S4     │ O      │
│ 13  │ I1     │ S5     │ O      │
│ 14  │ I2     │ S5     │ O      │
│ 15  │ I3     │ S5     │ O      │

MixedModelsSim.nlevels — Function

nlevels(nlev, tag='S')

Return a Vector{String} of tag followed by 1:nlev left-padded with zeros

Examples

julia> show(nlevels(10))
["S01", "S02", "S03", "S04", "S05", "S06", "S07", "S08", "S09", "S10"]

MixedModelsSim.pooled! — Function

pooled!(df::DataFrame, cols::Type=Union{AbstractString,Missing})

Like DataFrames.categorical! but converting columns to PooledArrays

MixedModelsSim.power_table — Function

power_table(sim, alpha = 0.05)

Returns a DataFrame with two columns, coefname and power, with the proportion of simulated p-values less than alpha, for sim, the output of simulate_waldtests.

MixedModelsSim.replicate — Method

replicate(f::Function, n::Integer; use_threads=false)

Return a vector of the values of n calls to f() - used in simulations where the value of f is stochastic.

Note that if f() is not thread-safe or depends on a non thread-safe RNG, then you must set use_threads=false. Also note that ordering of replications is not guaranteed when use_threads=true, although the replications are not otherwise affected for thread-safe f().

MixedModelsSim.sim_to_df — Method

sim_to_df(sim)

Returns a DataFrame with 6 columns: iteration, coefname, beta, se, z, p. Rows are all the coefficients for each iteration of sim, the output of simulate_waldtests. iteration is not guaranteed to be the same across runs of simulate_waldtests with the same seed, even though the samples will be.

MixedModelsSim.simdat_crossed — Function

simdat_crossed(RNG, subj_n, item_n; subj_btwn, item_btwn, both_win)
simdat_crossed(subj_n, item_n; subj_btwn, item_btwn, both_win)

Return a DataFrame with a design specified by the:

number of subjects (subj_n),

number of items (item_n)

between-subject factors (subj_btwn)

between-item factors (item_btwn)

within-subject/item factors (both_win)

Factors should be specified as Dicts in the following format:

Dict(
    "factor1_name" => ["F1_level1", "F1_level2"],
    "factor2_name" => ["F2_level1", "F2_level2", "F2_level3"]
)

MixedModelsSim.simulate_waldtests — Method

simulate_waldtests(rng::AbstractRNG, nsamp::Integer, m::LinearMixedModel;
    β = m.β, σ = m.σ, θ = m.θ, use_threads=false)
simulate_waldtests(nsamp::Integer, m::LinearMixedModel;
    β = m.β, σ = m.σ, θ = m.θ, use_threads=false)

Return a Vector of βs, z- and p-values for each coefficient in a mixed model. This is similar to the MixedModels.parametricbootstrap, but returns test statistics instead of estimates. This is useful for power analyses.

Perform nsamp parametric bootstrap replication fits of m, returning a Vector of named tuples of the associated z- and p-values for the coefficents/

The default random number generator is Random.GLOBAL_RNG.

Named Arguments

β, σ, and θ are the values of m's parameters for simulating the responses.

use_threads determines whether threads are used to parallelize the computation. Note that this functionality depends on the development version

using MixedModels, MixedModelsSim
using DataFrames, Gadfly, Random, StaticArrays, StatsBase, Tables

kb07 = MixedModels.dataset(:kb07);
form = @formula(rt_raw ~ 1 + spkr + prec + load + (1+spkr+prec+load|subj) + (1+spkr+prec+load|item));
cont = Dict(:spkr => HelmertCoding(),
            :prec => HelmertCoding(),
            :load => HelmertCoding())
fm1 = fit(MixedModel, form, kb07, contrasts=cont, REML=false);
zpmt = simulate_waldtests(MersenneTwister(42),10,fm1,use_threads=true);

mean(getindex.(columntable(zpmt).p, 1) .< 0.05)
mean(getindex.(columntable(zpmt).p, Symbol("(Intercept)")) .< 0.05)

mean(getindex.(columntable(zpmt).p, 2) .< 0.05)
mean(getindex.(columntable(zpmt).p, Symbol("spkr: old")) .< 0.05)

samples = DataFrame(columntable(zpmt).β)
shortestcovint(samples[!,Symbol("prec: maintain")])

plot(samples, layer(x=Symbol("load: yes")),
    Geom.density,
    Guide.xlabel("Bootstrap estimates of 'load: yes'"))

plot(stack(samples),
   layer(x=:value, color=:variable, xgroup=:variable),
   Geom.subplot_grid(Geom.density, free_x_axis=true),
   Guide.xlabel("Bootstrap estimates of β"))

MixedModelsSim.withinitem — Method

withinitem(nitem, df)

Return a DataFrame of item with nitem levels and balanced columns of conditions from df